requirements.txt
openai
black
gradio
python-dotenv